backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.5.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic[email]==2.5.3
reportlab==4.0.8
apscheduler==3.10.4
python-multipart==0.0.6
websockets==12.0
pydantic-settings==2.1.0
